backend/requirements.txt
# =============================================================================
# Core Framework
# =============================================================================
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic[email]>=2.6.0
pydantic-settings>=2.1.0
python-multipart>=0.0.9

# =============================================================================
# Database
# =============================================================================
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
pgvector>=0.2.0

# =============================================================================
# Caching & Task Queue
# =============================================================================
redis>=5.0.0
celery[redis]>=5.3.0
flower>=2.0.0

# =============================================================================
# Content Extraction
# =============================================================================
trafilatura>=1.6.0
httpx>=0.26.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
feedparser>=6.0.0

# =============================================================================
# AI/ML & Utilities
# =============================================================================
google-generativeai>=0.8.0
numpy>=1.26.0
scikit-learn>=1.4.0
python-dotenv>=1.0.0
structlog>=24.1.0
tenacity>=8.2.0
aiohttp>=3.9.0


# =============================================================================
# Security / Authentication
# =============================================================================
# =============================================================================
# Security / Authentication
# =============================================================================
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
